pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "max-assistant-v2"
version = "0.1.0"
description = "FRANK Assistant - Assistant IA vocal local modulaire avec orchestration, tools, HUD animé et RAG local."
authors = [
    { name = "Bruno Ahée" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv",
    "requests",
    "pydantic",
    "opencv-python",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/brunoah/max_assistant_v2"
Repository = "https://github.com/brunoah/max_assistant_v2"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]